requirements.txt
asgiref==3.4.1
boto3==1.18.11
botocore==1.21.11
Django==3.2.5
django-cors-headers==3.7.0
django-dotenv==1.4.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.3.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
